feat: add live vm validators to abe2e
