+++ e2e/nginx.yaml
@@ -1,0 +1,21 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  creationTimestamp: null
+  labels:
+    run: nginx
+  name: nginx
+spec:
+  hostNetwork: true
+  nodeSelector:
+    kubernetes.azure.com/agentpool: nodepool1
+  hostPID: true
+  containers:
+  - image: nginx
+    name: nginx
+    securityContext:
+      privileged: true
+      capabilities:
+        add: ["SYS_PTRACE", "SYS_RAWIO"]
+  dnsPolicy: ClusterFirst
+  restartPolicy: Never
