+++ e2e/template.go
@@ -1,7 +1,11 @@
 package e2e_test
 
 import (
+	"context"
 	"fmt"
+	"strconv"
+	"strings"
+	"testing"
 
 	"github.com/Azure/agentbaker/pkg/agent/datamodel"
 )
@@ -438,6 +442,39 @@ func baseTemplate() *datamodel.NodeBootstrappingConfiguration {
 		DisableUnattendedUpgrades: true,
 		SSHStatus:                 0,
 	}
+}
+
+func getBaseNodeBootstrappingConfiguration(ctx context.Context, t *testing.T, cloud *azureClient, suiteConfig *suiteConfig, clusterParams map[string]string) (*datamodel.NodeBootstrappingConfiguration, error) {
+	nbc := baseTemplate()
+	nbc.ContainerService.Properties.CertificateProfile.CaCertificate = clusterParams["/etc/kubernetes/certs/ca.crt"]
+
+	bootstrapKubeconfig := clusterParams["/var/lib/kubelet/bootstrap-kubeconfig"]
+
+	bootstrapToken, err := extractKeyValuePair("token", bootstrapKubeconfig)
+	if err != nil {
+		return nil, fmt.Errorf("failed to extract bootstrap token via regex: %q", err)
+	}
+
+	bootstrapToken, err = strconv.Unquote(bootstrapToken)
+	if err != nil {
+		return nil, fmt.Errorf("failed to unquote bootstrap token: %q", err)
+	}
+
+	server, err := extractKeyValuePair("server", bootstrapKubeconfig)
+	if err != nil {
+		return nil, fmt.Errorf("failed to extract fqdn via regex: %q", err)
+	}
+	tokens := strings.Split(server, ":")
+	if len(tokens) != 3 {
+		return nil, fmt.Errorf("expected 3 tokens from fqdn %q, got %d", server, len(tokens))
+	}
+	// strip off the // prefix from https://
+	fqdn := tokens[1][2:]
+
+	nbc.KubeletClientTLSBootstrapToken = &bootstrapToken
+	nbc.ContainerService.Properties.HostedMasterProfile.FQDN = fqdn
+
+	return nbc, nil
 }
 
 func getDebugDaemonset() string {
