+++ e2e/utils.go
@@ -1,0 +1,16 @@
+package e2e_test
+
+import "strings"
+
+func strToBoolMap(str string) map[string]bool {
+	str = strings.ReplaceAll(str, " ", "")
+	if str == "" {
+		return nil
+	}
+	parts := strings.SplitN(str, ",", -1)
+	m := make(map[string]bool, len(parts))
+	for _, p := range parts {
+		m[p] = true
+	}
+	return m
+}
